+++ test/run_all_versions.sh
@@ -228,6 +228,7 @@ SUITES=(
 	parallel_degree
 	parallel_export_parquet
 	parallel_flush_optin
+	parallel_scan_cost
 	parallel_vector_agg
 	parquet_count_bounds
 	parquet_export